Add module creditFooter.js to add a credit footer to the web page
The default exported funtion setCreditFooter(data) creates a new
<footer> element and append it to the page.

Then, it creates a .credit-footer div with the credit footer inside.
This div is then appended to the <footer>, and returned as well.
This function should be created after the structure of the webpage
has been created, in the end.

'data' is an optional parameter with whom you can specify a
different displayed name and github name, which will be used in
the credit footer. Since the function will be primarly be used for
my personal projects, the default data are my data.

The style of this credit footer is also applied by this function as
an inline style, so no further style is required.

The font family are imported using the webfontloader npm module
See: https://github.com/typekit/webfontloader